[#14] feat: ProjectFileManager 타입 생성 및 파일 업데이트 과정 추가
